pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "numen"
version = "0.1.0"
description = "AI-Augmented Terminal Notepad"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Aguiar", email = "notpublic@private.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]==0.15.2",
    "rich==13.6.0",
    "markdown-it-py==3.0.0",
    "fs==2.4.16",
    "pyyaml==6.0.1",
    "toml==0.10.2",
    "anthropic==0.21.3", 
    "openai==1.23.0",
    "requests==2.31.0",
    "python-frontmatter==1.1.0",
    "google-generativeai>=0.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.10.1",
    "mypy>=1.0.0",
]

[project.scripts]
numen = "numen.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 